Use single pip command to install packages
The existing method was using a single pip command for every package.
This could lead to errors, when the user try to fix a library A to a certain
version and another library B is also used, which referencing library B
(so this will use the latest allowed version of library A).
Now with the single call the pip command will use the correct versions.

Change-Id: I723f443597898df24bd714a57bc49fe3bcf091e6